Initial commit: Set up a development WordPress site with Docker containers.
This commit adds the necessary files for setting up a development WordPress website using Docker containers and provides steps for configuring and running the containers. It includes a MySQL database, WordPress, phpMyAdmin, and Nginx services, as well as a `nginx.conf` file that can be customized. The commit also includes a `README.md` file with instructions on how to set up and use the containers. The project is licensed under the MIT License.